Refactor indexation form
