+++ assets/customAgGridFunctions.js
@@ -10,6 +10,26 @@ dagcomponentfuncs.ScanLinkRenderer = function (props) {
     );
 };
 
+dagcomponentfuncs.PeakIndexLinkRenderer = function (props) {
+    // props.value will be the peakindex_id for the current row
+    const url = `/indexedpeak?indexid=${props.value}`;
+    return React.createElement(
+        'a',
+        { href: url },
+        props.value // This will be the text of the link (the peakindex_id)
+    );
+};
+
+dagcomponentfuncs.DatasetIdScanLinkRenderer = function (props) {
+    // props.value will be the dataset_id for the current row
+    const url = `/scan?scanid=${props.value}`;
+    return React.createElement(
+        'a',
+        { href: url },
+        props.value // This will be the text of the link (the dataset_id)
+    );
+};
+
 dagcomponentfuncs.ActionButtonsRenderer = function (props) {
     const { data } = props; // data contains the row data
