+++ laue_portal/pages/create_indexedpeaks.py
@@ -11,6 +11,7 @@ import laue_portal.database.db_schema as db_schema
 import laue_portal.components.navbar as navbar
 from sqlalchemy.orm import Session
 import laueindexing.pyLaueGo as pyLaueGo
+import laue_portal.components.peakindex_form as peakindex_form
 
 dash.register_page(__name__)
 
@@ -47,7 +48,7 @@ layout = dbc.Container(
             dbc.Button('Submit', id='submit_peakindex', color='primary'),
         ),
         html.Hr(),
-        ui_shared.peakindex_form,
+        peakindex_form,
     ],
     )
     ],
